refactored department pages
